third_party/blink/web_tests/svg/dynamic-updates/SVGFEDiffuseLightingElement-dom-kernelUnitLength-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'kernelUnitLength' attribute of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.getAttribute('kernelUnitLength') is "2"
PASS gradientElement.getAttribute('kernelUnitLength') is "3 4"
PASS successfullyParsed is true

TEST COMPLETE
